Migrate WalletConnect modals UX needs to be tested (should we close the first modal manually?)
